refactor(Rol): update import paths
