+++ src/pages/privileges/outlets/roles/index.tsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from "react";
 
 import { getData } from "@src/mocks/utils/dataMock.service";
+import { IRol } from "@pages/privileges/outlets/roles/add-role/forms/types/types";
 
 import { RolesUI } from "./interface";
-import { IRol } from "./types";
 
 export function Roles() {
   const [searchRole, setSearchRole] = useState<string>("");
